+++ back-end/routes/inspections.ts
@@ -1,0 +1,10 @@
+import { Router } from 'express';
+import * as inspectionsController from '../controllers/inspectionsController';
+
+const router = Router();
+
+router.get('/', inspectionsController.getAllInspections);
+router.post('/', inspectionsController.createInspection);
+
+export default router;
+
