+++ back-end/routes/inspectors.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import * as inspectorsController from '../controllers/inspectorsController';
+
+const router = Router();
+
+router.get('/', inspectorsController.getAllInspectors);
+
+export default router;
+
