+++ back-end/routes/properties.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import * as propertiesController from '../controllers/propertiesController';
+
+const router = Router();
+
+router.get('/', propertiesController.getAllProperties);
+
+export default router;
+
