+++ back-end/server.ts
@@ -6,8 +6,15 @@ import dotenv from 'dotenv';
 import swaggerJSDoc from 'swagger-jsdoc'; //specification generator
 import swaggerUI from 'swagger-ui-express'; //hosts GUI server
 import swaggerOptions from './swagger.config'; //configuration file that includes settings for the guide
-import { supabase, supabaseAdmin } from './supabase';
-import DatabaseService from './database';
+
+// Routes
+import healthRoutes from './routes/health';
+import organizationsRoutes from './routes/organizations';
+import propertiesRoutes from './routes/properties';
+import inspectionsRoutes from './routes/inspections';
+import inspectorsRoutes from './routes/inspectors';
+import observationsRoutes from './routes/observations';
+import supabaseRoutes from './routes/supabase';
 
 const app = express();
 const port = process.env.PORT || 4000;
@@ -35,194 +42,13 @@ const swaggerSpec = swaggerJSDoc(swaggerOptions);
 app.use('/api', swaggerUI.serve, swaggerUI.setup(swaggerSpec));
 
 // Routes
-app.get('/', (req, res) => {
-  res.json({ 
-    message: 'Backend server is running yayy!',
-    timestamp: new Date().toISOString()
-  });
-});
-
-app.get('/health', (req, res) => {
-  res.json({ status: 'healthy', service: 'back-end' });
-});
-
-// Supabase test route
-app.get('/supabase/test', async (req, res) => {
-  try {
-    const result = await DatabaseService.testConnection();
-    
-    if (!result.success) {
-      console.error('Supabase connection error:', result.error);
-      return res.status(500).json({ 
-        error: 'Supabase connection failed', 
-        details: result.error instanceof Error ? result.error.message : 'Unknown error'
-      });
-    }
-    
-    return res.json({ 
-      message: 'Supabase connection successful',
-      supabaseConnected: true,
-      timestamp: new Date().toISOString()
-    });
-  } catch (err) {
-    console.error('Supabase test error:', err);
-    return res.status(500).json({ 
-      error: 'Supabase test failed', 
-      details: err instanceof Error ? err.message : 'Unknown error'
-    });
-  }
-});
-
-// Organizations API
-app.get('/organizations', async (req, res) => {
-  try {
-    const { data, error } = await DatabaseService.fetchData('organizations');
-    
-    if (error) {
-      return res.status(500).json({ error: error instanceof Error ? error.message : 'Unknown error' });
-    }
-    
-    return res.json({ organizations: data });
-  } catch (err) {
-    console.error('Database query error:', err);
-    return res.status(500).json({ 
-      error: 'Database query failed', 
-      details: err instanceof Error ? err.message : 'Unknown error'
-    });
-  }
-});
-
-// Properties API
-app.get('/api/properties', async (req, res) => {
-  try {
-    const { organization_id } = req.query;
-    const filters = organization_id ? { organization_id: organization_id as string } : undefined;
-    
-    const { data, error } = await DatabaseService.fetchData('properties', '*', filters);
-    
-    if (error) {
-      return res.status(500).json({ error: error instanceof Error ? error.message : 'Unknown error' });
-    }
-    
-    return res.json({ properties: data });
-  } catch (err) {
-    console.error('Database query error:', err);
-    return res.status(500).json({ 
-      error: 'Database query failed', 
-      details: err instanceof Error ? err.message : 'Unknown error'
-    });
-  }
-});
-
-// Inspections API
-app.get('/api/inspections', async (req, res) => {
-  try {
-    const { organization_id, inspector_id, status } = req.query;
-    const filters: Record<string, any> = {};
-    
-    if (organization_id) filters.organization_id = organization_id as string;
-    if (inspector_id) filters.inspector_id = inspector_id as string;
-    if (status) filters.status = status as string;
-    
-    const { data, error } = await DatabaseService.fetchData('inspections', '*', Object.keys(filters).length ? filters : undefined);
-    
-    if (error) {
-      return res.status(500).json({ error: error instanceof Error ? error.message : 'Unknown error' });
-    }
-    
-    return res.json({ inspections: data });
-  } catch (err) {
-    console.error('Database query error:', err);
-    return res.status(500).json({ 
-      error: 'Database query failed', 
-      details: err instanceof Error ? err.message : 'Unknown error'
-    });
-  }
-});
-
-// Create new inspection
-app.post('/api/inspections', async (req, res) => {
-  try {
-    const { inspector_id, property_id, organization_id, status = 'draft', scheduled_for } = req.body;
-    
-    if (!inspector_id || !property_id || !organization_id) {
-      return res.status(400).json({ error: 'Missing required fields: inspector_id, property_id, organization_id' });
-    }
-    
-    const inspectionData = {
-      inspector_id,
-      property_id,
-      organization_id,
-      status,
-      scheduled_for: scheduled_for || null
-    };
-    
-    const { data, error } = await DatabaseService.insertData('inspections', inspectionData);
-    
-    if (error) {
-      return res.status(500).json({ error: error instanceof Error ? error.message : 'Unknown error' });
-    }
-    
-    return res.status(201).json({ inspection: data });
-  } catch (err) {
-    console.error('Database insert error:', err);
-    return res.status(500).json({ 
-      error: 'Database insert failed', 
-      details: err instanceof Error ? err.message : 'Unknown error'
-    });
-  }
-});
-
-// Inspectors API
-app.get('/api/inspectors', async (req, res) => {
-  try {
-    const { organization_id, active } = req.query;
-    const filters: Record<string, any> = {};
-    
-    if (organization_id) filters.organization_id = organization_id as string;
-    if (active !== undefined) filters.active = active === 'true';
-    
-    const { data, error } = await DatabaseService.fetchData('inspectors', '*', Object.keys(filters).length ? filters : undefined);
-    
-    if (error) {
-      return res.status(500).json({ error: error instanceof Error ? error.message : 'Unknown error' });
-    }
-    
-    return res.json({ inspectors: data });
-  } catch (err) {
-    console.error('Database query error:', err);
-    return res.status(500).json({ 
-      error: 'Database query failed', 
-      details: err instanceof Error ? err.message : 'Unknown error'
-    });
-  }
-});
-
-// Observations API
-app.get('/api/observations', async (req, res) => {
-  try {
-    const { section_id, severity, status } = req.query;
-    const filters: Record<string, any> = {};
-    
-    if (section_id) filters.section_id = section_id as string;
-    if (severity) filters.severity = severity as string;
-    if (status) filters.status = status as string;
-    
-    const { data, error } = await DatabaseService.fetchData('observations', '*', Object.keys(filters).length ? filters : undefined);
-    
-    if (error) {
-      return res.status(500).json({ error: error instanceof Error ? error.message : 'Unknown error' });
-    }
-    
-    return res.json({ observations: data });
-  } catch (err) {
-    console.error('Database query error:', err);
-    return res.status(500).json({ 
-      error: 'Database query failed', 
-      details: err instanceof Error ? err.message : 'Unknown error'
-    });
-  }
-});
+app.use('/', healthRoutes); // ✅ Health check routes (/, /health)
+app.use('/organizations', organizationsRoutes); // ✅ Organization-related routes
+app.use('/api/properties', propertiesRoutes); // ✅ Property-related routes
+app.use('/api/inspections', inspectionsRoutes); // ✅ Inspection-related routes
+app.use('/api/inspectors', inspectorsRoutes); // ✅ Inspector-related routes
+app.use('/api/observations', observationsRoutes); // ✅ Observation-related routes
+app.use('/supabase', supabaseRoutes); // ✅ Supabase test routes
 
 // Start server
 app.listen(port, () => {
